+++ src/data/const.js
@@ -3,3 +3,24 @@ export const FilterType = {
   FUTURE: `future`,
   PAST: `past`
 };
+
+export const Mode = {
+  DEFAULT: `default`,
+  EDIT: `edit`,
+  ADDING: `adding`
+};
+
+export const EmptyEvent = {
+  price: ``,
+  start: new Date().toISOString(),
+  end: new Date().toISOString(),
+  destination: {
+    description: ``,
+    name: ``,
+    pictures: []
+  },
+  id: 0,
+  isFavorite: false,
+  offers: [],
+  type: `flight`
+};
